requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
chromadb==0.3.21
numpy==1.23.2
openai==0.27.2
pandas==1.4.4
pypdf==3.8.1
python-dotenv==1.0.0
Requests==2.29.0
validators==0.20.0
langchain==0.0.154
urllib3==1.26.5